refactor: ArrowCardPair
# Conflicts:
#	packages/wallet-react-ui/src/signing/pages/CollectionApproval.tsx
#	packages/wallet-react-ui/src/signing/pages/Erc20Approval.tsx
#	packages/wallet-react-ui/src/signing/pages/Erc20Transfer.tsx
#	packages/wallet-react-ui/src/signing/pages/EthTransfer.tsx
